RCL-122 Add FOREGROUND Microphone permission and implement service on user call
