+++ app/build.gradle
@@ -12,8 +12,8 @@ android {
         minSdkVersion 21
         compileSdkVersion 34
         targetSdkVersion 34
-        versionCode 551
-        versionName "0.6.10.21"
+        versionCode 552
+        versionName "0.6.10.22"
         testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         vectorDrawables.useSupportLibrary = true
         multiDexEnabled true
