+++ app/src/main/AndroidManifest.xml
@@ -188,7 +188,13 @@
         <!--FIX me and add other foregroundServiceTypes below-->
         <service android:name=".CallService"
             android:enabled="true"
-            android:foregroundServiceType="phoneCall|microphone"
+            android:foregroundServiceType="phoneCall"
+            android:permission="android.permission.FOREGROUND_SERVICE"/>
+
+        <service
+            android:name=".CallStatusService"
+            android:exported="false"
+            android:foregroundServiceType="microphone"
             android:permission="android.permission.FOREGROUND_SERVICE"/>
 
         <service
