+++ app/src/main/kotlin/org/rivchain/cuplink/CallStatusService.kt
@@ -1,0 +1,83 @@
+package org.rivchain.cuplink
+
+import android.app.Notification
+import android.app.NotificationChannel
+import android.app.NotificationManager
+import android.app.PendingIntent
+import android.app.Service
+import android.content.Intent
+import android.content.pm.ServiceInfo
+import android.os.Build
+import android.os.IBinder
+import androidx.core.app.NotificationCompat
+
+class CallStatusService : Service() {
+
+    companion object {
+        const val CHANNEL_ID = "CallStatusChannel"
+        const val NOTIFICATION_ID = 101
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        createNotificationChannel()
+        val notification = buildNotification()
+        // Start the service as a foreground service with microphone permission for Android 10+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            startForeground(NOTIFICATION_ID, notification, ServiceInfo.FOREGROUND_SERVICE_TYPE_MICROPHONE)
+        } else {
+            // Start the service as a regular foreground service for older versions
+            startForeground(NOTIFICATION_ID, notification)
+        }
+    }
+
+    override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
+        // Handle service tasks here
+        return START_STICKY
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        stopForeground(true)
+    }
+
+    override fun onBind(intent: Intent?): IBinder? {
+        // This service is not designed for binding, return null
+        return null
+    }
+
+    private fun createNotificationChannel() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            val channel = NotificationChannel(
+                CHANNEL_ID,
+                "Call Status Service",
+                NotificationManager.IMPORTANCE_DEFAULT
+            )
+            val manager = getSystemService(NotificationManager::class.java)
+            manager?.createNotificationChannel(channel)
+        }
+    }
+
+    private fun buildNotification(): Notification {
+        // Create an Intent to open CallActivity when the user taps the notification
+        val notificationIntent = Intent(this, CallActivity::class.java)
+        // Use the FLAG_ACTIVITY_CLEAR_TOP and FLAG_ACTIVITY_SINGLE_TOP flags to bring the existing
+        // CallActivity to the foreground if it exists, instead of creating a new instance.
+        notificationIntent.flags = Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_SINGLE_TOP
+
+        // Create a PendingIntent that wraps the intent for launching CallActivity
+        val pendingIntent = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+            PendingIntent.getActivity(this, 0, notificationIntent, PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE)
+        } else {
+            PendingIntent.getActivity(this, 0, notificationIntent, PendingIntent.FLAG_UPDATE_CURRENT)
+        }
+        return NotificationCompat.Builder(this, CHANNEL_ID)
+            .setContentTitle("In call")
+            .setContentText("...")
+            .setSmallIcon(R.drawable.cup_link_small)
+            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
+            .setContentIntent(pendingIntent) // Attach the PendingIntent to the notification
+            .setAutoCancel(false) // Make sure notification is not dismissed when tapped
+            .build()
+    }
+}
